requirements.txt
fastapi
uvicorn
python-dotenv
supabase
langchain
langchain-openai
pydantic